+++ tests/unit/test_conf.py
@@ -1,7 +1,12 @@
 import contextlib
 
 import pytest
-from kibitzr.conf import settings, ReloadableSettings, ConfigurationError
+from kibitzr.conf import (
+    settings,
+    ReloadableSettings,
+    PlainYamlCreds,
+    ConfigurationError,
+)
 from ..compat import mock
 
 
@@ -78,22 +83,34 @@ notifiers = {
 }
 
 
-def test_complex_conf_sample():
-    with patch_source("open_conf", sample_conf):
-        with patch_source("open_creds", sample_creds):
-            conf = ReloadableSettings('::')
-    assert conf.checks == checks
-    assert conf.creds == creds
-
-
 @contextlib.contextmanager
-def patch_source(method, data):
+def patch_source(klass, method, data):
     fake_file = mock.mock_open(read_data=data)
-    with mock.patch.object(ReloadableSettings,
+    with mock.patch.object(klass,
                            method,
                            fake_file,
                            create=True) as fake_method:
         yield fake_method
+
+
+@contextlib.contextmanager
+def patch_creds(data):
+    with patch_source(PlainYamlCreds, 'open_creds', data) as fake_method:
+        yield fake_method
+
+
+@contextlib.contextmanager
+def patch_conf(data):
+    with patch_source(ReloadableSettings, 'open_conf', data) as fake_method:
+        yield fake_method
+
+
+def test_complex_conf_sample():
+    with patch_conf(sample_conf):
+        with patch_creds(sample_creds):
+            conf = ReloadableSettings('::')
+    assert conf.checks == checks
+    assert conf.creds['mailgun']['url'] == 'ZZZ'
 
 
 @mock.patch("kibitzr.conf.os.path.exists", return_value=False)
@@ -109,12 +126,12 @@ def test_reread():
         "    url: URL\n"
     )
     conf2 = conf1 + "    period: 60\n"
-    with patch_source("open_creds", ""):
-        with patch_source("open_conf", conf1):
+    with patch_creds(""):
+        with patch_conf(conf1):
             conf = ReloadableSettings('::')
             # Config didn't change:
             assert not conf.reread()
-        with patch_source("open_conf", conf2):
+        with patch_conf(conf2):
             # Config changed:
             assert conf.reread()
     assert conf.checks == [{'name': 'Name', 'url': 'URL', 'period': 60}]
@@ -122,14 +139,14 @@ def test_reread():
 
 def test_name_from_url_population():
     conf = "checks: [{url: 'http://example.com/page_name'}]"
-    with patch_source("open_conf", conf):
+    with patch_conf(conf):
         conf = ReloadableSettings('::')
     assert conf.checks[0]['name'] == 'http-example-com-page_name'
 
 
 def test_unnamed_check():
     conf = "checks: [{period: 1}]"
-    with patch_source("open_conf", conf):
+    with patch_conf(conf):
         conf = ReloadableSettings('::')
     assert conf.checks == [{
         'name': 'Unnamed check 1',
@@ -139,13 +156,13 @@ def test_unnamed_check():
 
 def test_period_parse():
     conf = "checks: [{period: 1 hour}]"
-    with patch_source("open_conf", conf):
+    with patch_conf(conf):
         conf = ReloadableSettings('::')
     assert conf.checks[0]['period'] == 3600
 
 
 def test_empty_period():
     conf = "checks: [{name: x}]"
-    with patch_source("open_conf", conf):
+    with patch_conf(conf):
         conf = ReloadableSettings('::')
     assert conf.checks[0]['period'] == 300